requirements.txt
discord.py
aiohttp
ytmusicapi
requests
flask
